+++ custom_model_runner/CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### [1.4.15] - 2021-02-11
+#### Fixed
+- dataset temp naming for failed validation checks
+- command line argument parsing
+
 #### [1.4.14] - 2021-02-08
 #### Added
 - make DRUM return predictions in DR format when deployment config is provided
 ##### Changed
 - default perf tests timeout from 180 to 600 s
 - refactor model metadata YAML schema
-
 
 #### [1.4.13] - 2021-02-01
 #### Added